[XRay] Fix FunctionRecord serialization
This change makes the writer implementation more consistent with the way
fields are written down to avoid assumptions on bitfield order and
padding. We also fix an inconsistency between the type returned by the
`delta()` accessor to match the data member it's returning.

This is a follow-up to D51289 and D51210.

llvm-svn: 341230

+++ llvm/include/llvm/XRay/FDRRecords.h
@@ -244,7 +244,7 @@ public:
   // properties.
   RecordTypes recordType() const { return Kind; }
   int32_t functionId() const { return FuncId; }
-  uint64_t delta() const { return Delta; }
+  uint32_t delta() const { return Delta; }
 
   Error apply(RecordVisitor &V) override;
 };
